+++ extensions/gmail/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Gmail Changelog
 
+## [Labels] - 2023-08-10
+
+- Add Labels support via actions
+
 ## [Initial Version] - 2023-08-09